Extract serializer to separate package
